loggedIn user can only delete their own comments

+++ src/Components/Article_one/CommentCard.jsx
@@ -7,9 +7,12 @@ function CommentCard(props) {
   const { body, votes, comment_id, author } = comment;
 
   const eraseComment = () => {
+    console.log(author)
+    if (author === "jessjelly"){
     handleDelete(comment_id).then(displayDeleted(comment_id)).catch(err => {
       console.dir(err);
-    })
+    })}
+    else{alert("You can only delete your OWN comments!")}
   };
 
   return (
